+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 Bugfixes:
 - Backbone Collections throw error in loadData (#606)
+- auto column size did not consider CSS style of each instance separately
 
 ## [0.8.22](https://github.com/warpech/jquery-handsontable/tree/v0.8.22) (Apr 29, 2013)